Sunil/requirements.txt
# Core Google AI Platform SDK
google-cloud-aiplatform==1.43.0

# Kubeflow Pipelines SDK
kfp==2.5.0
kfp-pipeline-spec==0.2.2
kfp-server-api==2.0.5

# Logging
python-json-logger==2.0.7

# Data handling
pandas==2.3.2
PyYAML==6.0.2
click==8.2.1
tabulate==0.9.0

# Networking
urllib3==1.26.20
requests-toolbelt==0.10.1
websocket-client==1.8.0

# Kubernetes client
kubernetes==26.1.0

# ✅ Pin protobuf to a version compatible with both TF 2.17.0 and KFP
protobuf==3.20.3
